學員作品/三大法人股票買賣超觀察_陳定康_江榮展_web/requirements.txt
python-dotenv>=1.0.0
pandas
numpy>=1.24.0
matplotlib>=3.7.0
tk>=0.1.0

finlab>=0.3.0
requests>=2.31.0
pyquery>=2.0.0
tqdm>=4.65.0
scipy>=1.11.0
openpyxl>=3.1.0

jupyter>=1.0.0
ipython>=8.12.0

dash
plotly
gunicorn==21.2.0
